requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
appnope==0.1.2
    # via
    #   ipykernel
    #   ipython
argon2-cffi==21.3.0
    # via notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==2.0.5
    # via stack-data
attrs==21.4.0
    # via jsonschema
backcall==0.2.0
    # via ipython
bleach==4.1.0
    # via nbconvert
cachetools==5.0.0
    # via google-auth
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via argon2-cffi-bindings
charset-normalizer==2.0.12
    # via requests
cycler==0.11.0
    # via matplotlib
db-dtypes==0.3.1
    # via pandas-gbq
debugpy==1.5.1
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
entrypoints==0.4
    # via
    #   jupyter-client
    #   nbconvert
et-xmlfile==1.1.0
    # via openpyxl
executing==0.8.3
    # via stack-data
fonttools==4.29.1
    # via matplotlib
gcloud==0.18.3
    # via -r requirements.in
google-api-core[grpc]==2.7.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   pandas-gbq
google-auth==2.6.0
    # via
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-core
    #   pandas-gbq
    #   pydata-google-auth
google-auth-oauthlib==0.5.0
    # via
    #   pandas-gbq
    #   pydata-google-auth
google-cloud==0.34.0
    # via -r requirements.in
google-cloud-bigquery==2.34.2
    # via
    #   -r requirements.in
    #   pandas-gbq
google-cloud-bigquery-storage==2.12.0
    # via pandas-gbq
google-cloud-core==2.2.3
    # via google-cloud-bigquery
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.3.2
    # via google-cloud-bigquery
googleapis-common-protos==1.55.0
    # via
    #   gcloud
    #   google-api-core
    #   grpcio-status
grpcio==1.44.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   grpcio-status
grpcio-status==1.44.0
    # via google-api-core
httplib2==0.20.4
    # via
    #   gcloud
    #   oauth2client
idna==3.3
    # via requests
importlib==1.0.4
    # via -r requirements.in
ipykernel==6.9.1
    # via
    #   ipywidgets
    #   notebook
ipython==8.1.1
    # via
    #   ipykernel
    #   ipywidgets
ipython-genutils==0.2.0
    # via
    #   ipywidgets
    #   nbformat
    #   notebook
ipywidgets==7.6.5
    # via -r requirements.in
jedi==0.18.1
    # via ipython
jinja2==3.0.3
    # via
    #   nbconvert
    #   notebook
jsonschema==4.4.0
    # via nbformat
jupyter-client==7.1.2
    # via
    #   ipykernel
    #   nbclient
    #   notebook
jupyter-core==4.9.2
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-widgets==1.0.2
    # via ipywidgets
kiwisolver==1.3.2
    # via matplotlib
markupsafe==2.1.0
    # via jinja2
matplotlib==3.5.1
    # via
    #   -r requirements.in
    #   seaborn
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
nbclient==0.5.12
    # via nbconvert
nbconvert==6.4.2
    # via notebook
nbformat==5.1.3
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.4
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
notebook==6.4.8
    # via widgetsnbextension
numpy==1.22.3
    # via
    #   -r requirements.in
    #   db-dtypes
    #   matplotlib
    #   pandas
    #   pandas-gbq
    #   patsy
    #   pyarrow
    #   scipy
    #   seaborn
    #   statsmodels
oauth2client==4.1.3
    # via gcloud
oauthlib==3.2.0
    # via requests-oauthlib
openpyxl==3.0.9
    # via -r requirements.in
packaging==21.3
    # via
    #   bleach
    #   db-dtypes
    #   google-cloud-bigquery
    #   matplotlib
    #   statsmodels
pandas==1.4.1
    # via
    #   -r requirements.in
    #   db-dtypes
    #   pandas-gbq
    #   seaborn
    #   statsmodels
pandas-gbq==0.17.3
    # via -r requirements.in
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
patsy==0.5.2
    # via statsmodels
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.0.1
    # via matplotlib
plotly==5.6.0
    # via -r requirements.in
prometheus-client==0.13.1
    # via notebook
prompt-toolkit==3.0.28
    # via ipython
proto-plus==1.20.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
protobuf==3.19.4
    # via
    #   gcloud
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pyarrow==6.0.1
    # via
    #   db-dtypes
    #   pandas-gbq
pyasn1==0.4.8
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via
    #   google-auth
    #   oauth2client
pycparser==2.21
    # via cffi
pydata-google-auth==1.3.0
    # via pandas-gbq
pygments==2.11.2
    # via
    #   ipython
    #   jupyterlab-pygments
    #   nbconvert
pyparsing==3.0.7
    # via
    #   httplib2
    #   matplotlib
    #   packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   google-cloud-bigquery
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2021.3
    # via pandas
pyzmq==22.3.0
    # via
    #   jupyter-client
    #   notebook
requests==2.27.1
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.8
    # via
    #   google-auth
    #   oauth2client
scipy==1.8.0
    # via
    #   seaborn
    #   statsmodels
seaborn==0.11.2
    # via -r requirements.in
send2trash==1.8.0
    # via notebook
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   gcloud
    #   google-auth
    #   grpcio
    #   oauth2client
    #   patsy
    #   plotly
    #   python-dateutil
stack-data==0.2.0
    # via ipython
statsmodels==0.13.2
    # via -r requirements.in
tenacity==8.0.1
    # via plotly
terminado==0.13.3
    # via notebook
testpath==0.6.0
    # via nbconvert
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
traitlets==5.1.1
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
urllib3==1.26.8
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
widgetsnbextension==3.5.2
    # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# setuptools
